// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        Status: {
// ...
            card?: {
                /** @description Description of preview */
                description: string;
                /**
                 * Format: uri
                 * @description Preview thumbnail
                 */
                image?: string | null;
                /** @description The provider of the original resource */
                provider_name?: string | null;
                /**
                 * Format: uri
                 * @description A link to the provider of the original resource
                 */
                provider_url?: string;
                /** @description Title of linked resource */
                title: string;
                /**
                 * @description The type of the preview card
                 * @enum {string}
                 */
                type: "link" | "photo" | "video" | "rich";
                /**
                 * Format: uri
                 * @description Location of linked resource
                 */
                url: string;
            } | null;
// ...
        };
// ...
    };
    responses: never;
    parameters: {
        /**
         * @description Account ID or nickname
         * @example 123
         */
        accountIdOrNickname: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}